[Common] Move the time.h and measure.h to common.
